feat/add joi validation error handler
add joi library validation error handler for first validate user input from body and give error direct to the error handler middleware and then goes to next function where all logic is written so there will other validations check like checking credential and give error , find user from email id and give error if not found , check user already exist or not in database by email id on registration page
