+++ auth/authRoutes.js
@@ -1,4 +1,5 @@
 import express from "express";
+import joiSchema from "../schemas/joiSchema.js";
 
 //controller files
 import registerController from "./register/registerController.js";
@@ -14,7 +15,7 @@ import customLoginMiddleware from "./middleware/customLoginMiddleware.js";
 
 const router = express.Router();
 
-//custom validation for register
+//custom validation
 router.post(
   "/register/custom-validation",
   customRegisterMiddleware.customRegisterValidation,
@@ -24,6 +25,19 @@ router.post(
 router.post(
   "/login/custom-validation",
   customLoginMiddleware.customLoginValidation,
+  loginController.loginUser
+);
+
+//joi validation
+router.post(
+  "/register/joi-validation",
+  authMiddleware.validate(joiSchema.userRegisterValidateSchema),
+  registerController.registerUser
+);
+
+router.post(
+  "/login/joi-validation",
+  authMiddleware.validate(joiSchema.userLoginValidateSchema),
   loginController.loginUser
 );
